Added image upload functionality.

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,23 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 class PostsController extends Controller
 {
+
+
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -36,21 +49,38 @@ class PostsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+  public function store(Request $request)
     {
-        $this->validate($request,[
-            'title'=>'required',
-            'body'=>'required'
+        $this->validate($request, [
+            'title' => 'required',
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
-        $post= new Post;
-
-        $post->title=$request->input('title');
-        $post->body=$request->input('body');
-        $post->user_id=auth()->user()->id;
+       
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+        // Create Post
+        $post = new Post;
+        $post->title = $request->input('title');
+        $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
-
-        return redirect('/posts')->with('success','Post is sucessfully published');
+        return redirect('/posts')->with('success', 'Post Created');
     }
 
     /**
@@ -93,9 +123,26 @@ class PostsController extends Controller
             'body'=>'required'
         ]);
 
+          // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }  
+
       
         $post->title=$request->input('title');
         $post->body=$request->input('body');
+         if($request->hasFile('cover_image')){
+        $post->cover_image=$fileNameToStore;    
+         }
         
         $post->save();
 
@@ -110,7 +157,15 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        $post= Post::find($id);
+          $post= Post::find($id);
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorised Page');
+        }
+      
+
+        if($post->cover_image !='noimage.jpg'){
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
 
         $post->delete();
         return redirect('/posts')->with('success','Post is sucessfully deleted');
